Fixed FutureOr<dynamic> error during compilation time, Used await instead of .then method for the ImageGallerySaver Function to remove the no .then function error,Fixed InlineSpan Error by replacing .text.text with .text.toPlainText()

+++ mp_chart/lib/mp/chart/chart.dart
@@ -46,15 +46,25 @@ abstract class ChartState<T extends Chart> extends State<T> {
 
     String fileName = DateTime.now().toIso8601String();
     String path = '$directory/$fileName.png';
-    _screenshotController.capture(path: path, pixelRatio: 3.0).then((imgFile) {
-      ImageGallerySaver.saveImage(Uint8List.fromList(imgFile.readAsBytesSync()))
-          .then((value) {
-        imgFile.delete();
-      });
+    var value;
+
+    var imgFile = await _screenshotController.capture(path: path, pixelRatio: 3.0);
+    if(imgFile != null ){
+      value = ImageGallerySaver.saveImage(Uint8List.fromList(imgFile.readAsBytesSync()));
+      if(value != null) imgFile.delete();
       isCapturing = false;
-    }).catchError((error) {
+    }else{
       isCapturing = false;
-    });
+    }
+    // _screenshotController.capture(path: path, pixelRatio: 3.0).then((imgFile) {
+    //   ImageGallerySaver.saveImage(Uint8List.fromList(imgFile.readAsBytesSync()))
+    //       .then((value) {
+    //     imgFile.delete();
+    //   });
+    //   isCapturing = false;
+    // }).catchError((error) {
+    //   isCapturing = false;
+    // });
   }
 
   @override
